auth_service/requirements.txt
Flask==3.0.0
bcrypt==4.1.2
python-dotenv==1.0.1
pymongo==4.6.1
python-dotenv==1.0.1
requests